liked porjects on employee deshboard

+++ gallery/backend/src/main/java/idm3/project/gallery/repository/ProjectRepository.java
@@ -1,17 +1,16 @@
 package idm3.project.gallery.repository;
 
 import idm3.project.gallery.model.Project;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
 import idm3.project.gallery.model.User;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ProjectRepository extends CrudRepository<Project, Long> {
+public interface ProjectRepository extends CrudRepository<Project, Integer> {
 
     @Query("SELECT DISTINCT p FROM Project p " +
             "JOIN UserProjectLike upl ON upl.project = p " +
@@ -19,4 +18,3 @@ public interface ProjectRepository extends CrudRepository<Project, Long> {
             "ORDER BY p.created DESC")
     List<Project> findProjectsLikedByUserOrderByCreatedDesc(@Param("user") User user);
 }
-
